Replace Transcript shim with adaptor, add String and Number adaptors
The Transcript shim wasn't as useful as I hoped, since the JavaScript
console's API is different. Instead, implement an adaptor that puts a
buffer in front of the console, and flushes it when a newline is written.
Not perfect, but better.

Also subclass the native String and Number types and enrich them to respond
to translated binary messages.

The above changes are not yet automatically incorporated into the output.
This needs some thought.