Execute generated code during unit tests
Change the test framework so that after generating and comparing the
JavaScript, it then generates it again with the runtime and executes it. It
expects the code to run without errors. This caught an issue with how
return constructs were being executed, and necessitated the definition of
the stReturn function.